refactor : common에서의 dto 패키지 설정
