+++ src/main/java/inu/codin/codin/common/security/dto/PortalLoginResponseDto.java
@@ -1,6 +1,6 @@
 package inu.codin.codin.common.security.dto;
 
-import inu.codin.codin.common.Department;
+import inu.codin.codin.common.dto.Department;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
