+++ src/main/java/inu/codin/codin/domain/info/domain/lab/entity/Lab.java
@@ -2,7 +2,7 @@ package inu.codin.codin.domain.info.domain.lab.entity;
 
 import inu.codin.codin.domain.info.domain.lab.dto.request.LabCreateUpdateRequestDto;
 import inu.codin.codin.domain.info.entity.Info;
-import inu.codin.codin.common.Department;
+import inu.codin.codin.common.dto.Department;
 import inu.codin.codin.domain.info.entity.InfoType;
 import jakarta.validation.constraints.NotBlank;
 import lombok.AccessLevel;
