+++ src/main/java/inu/codin/codin/domain/info/domain/office/dto/response/OfficeDetailsResponseDto.java
@@ -1,12 +1,11 @@
 package inu.codin.codin.domain.info.domain.office.dto.response;
 
-import inu.codin.codin.common.Department;
+import inu.codin.codin.common.dto.Department;
 import inu.codin.codin.domain.info.domain.office.entity.Office;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.Setter;
 
 import java.util.List;
