+++ src/main/java/inu/codin/codin/domain/info/domain/professor/entity/Professor.java
@@ -1,6 +1,6 @@
 package inu.codin.codin.domain.info.domain.professor.entity;
 
-import inu.codin.codin.common.Department;
+import inu.codin.codin.common.dto.Department;
 import inu.codin.codin.domain.info.domain.professor.dto.request.ProfessorCreateUpdateRequestDto;
 import inu.codin.codin.domain.info.entity.Info;
 import inu.codin.codin.domain.info.entity.InfoType;
