+++ src/main/java/inu/codin/codin/domain/info/domain/professor/service/ProfessorService.java
@@ -1,6 +1,6 @@
 package inu.codin.codin.domain.info.domain.professor.service;
 
-import inu.codin.codin.common.Department;
+import inu.codin.codin.common.dto.Department;
 import inu.codin.codin.domain.info.domain.professor.dto.response.ProfessorListResponseDto;
 import inu.codin.codin.domain.info.domain.professor.dto.response.ProfessorThumbnailResponseDto;
 import inu.codin.codin.domain.info.domain.professor.dto.request.ProfessorCreateUpdateRequestDto;
