+++ src/main/java/inu/codin/codin/domain/post/domain/comment/entity/CommentEntity.java
@@ -1,6 +1,6 @@
 package inu.codin.codin.domain.post.domain.comment.entity;
 
-import inu.codin.codin.common.BaseTimeEntity;
+import inu.codin.codin.common.dto.BaseTimeEntity;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
 import lombok.Getter;
