+++ src/main/java/inu/codin/codin/domain/report/entity/ReportEntity.java
@@ -1,7 +1,7 @@
 package inu.codin.codin.domain.report.entity;
 
 
-import inu.codin.codin.common.BaseTimeEntity;
+import inu.codin.codin.common.dto.BaseTimeEntity;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
 import lombok.Getter;
