+++ src/main/java/inu/codin/codin/infra/fcm/entity/FcmTokenEntity.java
@@ -1,6 +1,6 @@
 package inu.codin.codin.infra.fcm.entity;
 
-import inu.codin.codin.common.BaseTimeEntity;
+import inu.codin.codin.common.dto.BaseTimeEntity;
 import inu.codin.codin.domain.user.entity.UserEntity;
 import inu.codin.codin.infra.fcm.exception.FcmDuplicatedTokenException;
 import jakarta.validation.constraints.NotBlank;
